testing: Add script to test all apps builds
The `./build_all.sh` script will test the build for all the applications that
have scripts created in the `make-based/` directory, with certain PRs
and dependencies pulled.

To use it, run `./build_all.sh <PR number | branch> [deps]`.
The `PR number` is the number of the unikraft core pull request you want
to test. The `branch` argument is the unikraft core branch you want to
add the dependencies on top of (likely `staging` or `stable`).

The `deps` argument is a list of dependencies, in form of pull requests,
from the unikraft core or external libraries repositories.
For example, `./build_all.sh 123 musl/10 lwip/20 unikraft/200` will pull
PR 123 from the unikraft core, rebase it on top of PR 200 and add the
wanted PRs to the external libraries `musl` and `lwip`.

The results of the script will be shown as below, and the errors will be
saved in log files `err.log.<app-name>`:
```
Building nginx... FAILED
Building elfloader... PASSED
Building helloworld... PASSED
Building httpreply... PASSED
```

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>