make-based: Reset to release tag when doing setup
This feature updates the `setup` command to reset a repositories to a
release tag. This makes it easier to have a functional setup.
Irrespective of updates to different upstream repositories, it will will
provide a stable environment for configuring and building.

When running the `setup` command for each `do.sh` script, application,
libraries and Unikraft repositories will be checked out to a branch
created from the release tag. Release tags such as `RELEASE-0.13.1` will
result in the creation of a `br-RELEASE-0.13.1` branch in each
repository.

The tag is defined in the `include/base` file. Currently, there is no
per-release tag configuration file (for each app in the `files/.config`
file) ; this is something to be added later.

As of this commit, app configuration files (in `files/.config`) work for
the `RELEASE-0.13.1` tag.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>